Set up MD3 experiments pages
